rename root package to amlb
