+++ examples/custom/Decision_Tree/exec.py
@@ -1,11 +1,11 @@
 import logging
 from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
 
-from automl.benchmark import TaskConfig
-from automl.data import Dataset
-from automl.datautils import impute
-from automl.results import save_predictions_to_file
-from automl.utils import Timer
+from amlb.benchmark import TaskConfig
+from amlb.data import Dataset
+from amlb.datautils import impute
+from amlb.results import save_predictions_to_file
+from amlb.utils import Timer
 
 log = logging.getLogger(__name__)
