+++ examples/custom/RandomForest_standalone/__init__.py
@@ -1,4 +1,4 @@
-from automl.utils import call_script_in_same_dir
+from amlb.utils import call_script_in_same_dir
 
 
 def setup(*args):
