+++ frameworks/RandomForest/exec.py
@@ -8,11 +8,11 @@ os.environ['OPENBLAS_NUM_THREADS'] = '1'
 os.environ['MKL_NUM_THREADS'] = '1'
 from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 
-from automl.benchmark import TaskConfig
-from automl.data import Dataset
-from automl.datautils import impute
-from automl.results import save_predictions_to_file
-from automl.utils import Timer
+from amlb.benchmark import TaskConfig
+from amlb.data import Dataset
+from amlb.datautils import impute
+from amlb.results import save_predictions_to_file
+from amlb.utils import Timer
 
 log = logging.getLogger(__name__)
