+++ frameworks/TPOT/exec.py
@@ -12,11 +12,11 @@ os.environ['OPENBLAS_NUM_THREADS'] = '1'
 os.environ['MKL_NUM_THREADS'] = '1'
 from tpot import TPOTClassifier, TPOTRegressor
 
-from automl.benchmark import TaskConfig
-from automl.data import Dataset
-from automl.datautils import Encoder, impute, write_csv
-from automl.results import save_predictions_to_file
-from automl.utils import Timer, split_path, path_from_split
+from amlb.benchmark import TaskConfig
+from amlb.data import Dataset
+from amlb.datautils import Encoder, impute, write_csv
+from amlb.results import save_predictions_to_file
+from amlb.utils import Timer, split_path, path_from_split
 
 
 log = logging.getLogger(__name__)
