+++ frameworks/constantpredictor/exec.py
@@ -2,10 +2,10 @@ import logging
 
 from sklearn.dummy import DummyClassifier, DummyRegressor
 
-from automl.benchmark import TaskConfig
-from automl.data import Dataset
-from automl.results import save_predictions_to_file
-from automl.utils import Timer
+from amlb.benchmark import TaskConfig
+from amlb.data import Dataset
+from amlb.results import save_predictions_to_file
+from amlb.utils import Timer
 
 
 log = logging.getLogger(__name__)
