+++ frameworks/oboe/__init__.py
@@ -1,4 +1,4 @@
-from automl.utils import call_script_in_same_dir, dir_of
+from amlb.utils import call_script_in_same_dir, dir_of
 
 
 def setup(*args):
