+++ recover_results.py
@@ -2,12 +2,12 @@ import argparse
 import os
 
 # prevent asap other modules from defining the root logger using basicConfig
-import automl.logger
+import amlb.logger
 
 
-import automl
-from automl import log
-from automl.utils import Namespace as ns, config_load
+import amlb
+from amlb import log
+from amlb.utils import Namespace as ns, config_load
 
 
 parser = argparse.ArgumentParser()
@@ -20,12 +20,12 @@ extras = {t[0]: t[1] if len(t) > 1 else True for t in [x.split('=', 1) for x in 
 # log_dir = os.path.join(args.outdir if args.outdir else '.', 'logs')
 # os.makedirs(log_dir, exist_ok=True)
 # now_str = datetime_iso(date_sep='', time_sep='')
-automl.logger.setup(root_level='DEBUG', console_level='INFO')
+amlb.logger.setup(root_level='DEBUG', console_level='INFO')
 
 config = config_load("resources/config.yaml")
 config.run_mode = 'script'
 config.script = os.path.basename(__file__)
-automl.resources.from_configs(config, ns.parse(extras))
+amlb.resources.from_configs(config, ns.parse(extras))
 
-aws = automl.AWSBenchmark.fetch_results(args.instances, force_update=True)
+aws = amlb.AWSBenchmark.fetch_results(args.instances, force_update=True)
