+++ runscores.py
@@ -2,13 +2,13 @@ import argparse
 import os
 
 # prevent asap other modules from defining the root logger using basicConfig
-import automl.logger
+import amlb.logger
 
 from ruamel import yaml
 
-import automl
-from automl import log
-from automl.utils import config_load
+import amlb
+from amlb import log
+from amlb.utils import config_load
 
 
 parser = argparse.ArgumentParser()
@@ -20,13 +20,13 @@ args = parser.parse_args()
 # log_dir = os.path.join(args.outdir if args.outdir else '.', 'logs')
 # os.makedirs(log_dir, exist_ok=True)
 # now_str = datetime_iso(date_sep='', time_sep='')
-automl.logger.setup(root_level='DEBUG', console_level='INFO')
+amlb.logger.setup(root_level='DEBUG', console_level='INFO')
 
 config = config_load("resources/config.yaml")
 config.run_mode = 'script'
 config.script = os.path.basename(__file__)
-automl.resources.from_config(config)
+amlb.resources.from_config(config)
 
-scores = automl.TaskResult.score_from_predictions_file(args.predictions)
+scores = amlb.TaskResult.score_from_predictions_file(args.predictions)
 log.info("\n\nScores computed from %s:\n%s", args.predictions, yaml.dump(dict(scores), default_flow_style=False))
